Refactoring Sync by structuring the files
