+++ renderer/src/components/LoadingSpinner/index.js
@@ -1,0 +1,1 @@
+export { default as LoadingSpinner } from './LoadingSpinner';
