+++ renderer/src/components/ProgressCircle/ProgressCircle.js
