+++ renderer/src/components/ProgressCircle/index.js
@@ -1,0 +1,1 @@
+export { default as ProgressCircle } from './ProgressCircle';
