+++ renderer/src/components/Sync/Gitea/EditorSync/EditorSync.js
@@ -13,7 +13,7 @@ import PopUpModal from '../../PopUpModal';
 import * as logger from '../../../../logger';
 import CloudUploadIcon from '@/icons/basil/Outline/Files/Cloud-upload.svg';
 import CloudCheckIcon from '@/icons/basil/Solid/Files/Cloud-check.svg';
-import ProgressCircle from '../../ProgressCircle';
+import { ProgressCircle } from '../../../ProgressCircle';
 import Door43Logo from '@/icons/door43.svg';
 import { getGiteaUsersList, handleEditorSync } from './EditorSyncUtils';
 import useGetCurrentProjectMeta from '../../hooks/useGetCurrentProjectMeta';
