--- renderer/src/components/Sync/supportForExistingSyncUsers.js
@@ -1,3 +1,0 @@
-export const supportForExistingSyncUsers = async () => {
-// cupport codes checka and resolve all
-};
